+++ CHANGELOG.md
@@ -11,7 +11,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Language localization for Portuguese
 
+### Fixed
+
+- Missing localization keys in the `Vocabulary` interface
+- Missing translations for certain localization keys in `de.json` and `nl.json` to ensure all language files are complete
+
 ## [4.8.1] - 2026-02-09
+
+### Added
+
+- New `LocalizationConfig.adaptLocalizationToSubtitleLanguage` option to enable dynamically changing the UI language based on the user's selected subtitle language.
+- New `UIConfig.localization` config option to pass a `LocalizationConfig` within the `UIConfig` to the UI.
+
+### Changed
+
+- `ListItem.label` is now a `LocalizableText` instead of a `string`. It is strongly recommended to check behavior for any `ListSelectorConfig.filter` and `ListSelectorConfig.translator` implementations that might be in place.
 
 ### Fixed